Move TextEmbeddingIndexSettings from importing to vectorbase
