+++ python/ta/typeagent/aitools/vectorbase.py
@@ -1,13 +1,40 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 
+from dataclasses import dataclass
 from typing import Any, NamedTuple, TypedDict, cast
 
 import numpy as np
 from numpy.typing import NDArray
 
 from .embeddings import AsyncEmbeddingModel, NormalizedEmbedding, NormalizedEmbeddings
-from ..knowpro.importing import TextEmbeddingIndexSettings
+
+
+@dataclass
+class TextEmbeddingIndexSettings:
+    embedding_model: AsyncEmbeddingModel | None = None
+    embedding_size: int | None = None
+    min_score: float = 0.0
+    max_matches: int | None = None
+    retry_max_attempts: int = 2
+    retry_delay: float = 2.0  # Seconds
+    batch_size: int = 8
+
+    def __init__(
+        self,
+        embedding_model: AsyncEmbeddingModel | None = None,
+        embedding_size: int | None = None,
+        min_score: float | None = None,
+        max_matches: int | None = None,
+    ):
+        if embedding_model is None:
+            embedding_model = AsyncEmbeddingModel()
+        self.embedding_model = embedding_model
+        self.embedding_size = embedding_size
+        if min_score is None:
+            min_score = 0.85
+        self.min_score = min_score
+        self.max_matches = max_matches
 
 
 class ScoredOrdinal(NamedTuple):
