+++ python/ta/typeagent/knowpro/importing.py
@@ -4,35 +4,7 @@
 from dataclasses import dataclass, field
 
 from ..aitools.embeddings import AsyncEmbeddingModel
-
-
-# TODO: TextEmbeddingIndexSettings belongs in vectorbase.py.
-@dataclass
-class TextEmbeddingIndexSettings:
-    embedding_model: AsyncEmbeddingModel | None = None
-    embedding_size: int | None = None
-    min_score: float = 0.0
-    max_matches: int | None = None
-    retry_max_attempts: int = 2
-    retry_delay: float = 2.0  # Seconds
-    batch_size: int = 8
-
-    # TODO: Add embedding_model, embedding_size arguments (one supported).
-    def __init__(
-        self,
-        embedding_model: AsyncEmbeddingModel | None = None,
-        embedding_size: int | None = None,
-        min_score: float | None = None,
-        max_matches: int | None = None,
-    ):
-        if embedding_model is None:
-            embedding_model = AsyncEmbeddingModel()
-        self.embedding_model = embedding_model
-        self.embedding_size = embedding_size
-        if min_score is None:
-            min_score = 0.85
-        self.min_score = min_score
-        self.max_matches = max_matches
+from ..aitools.vectorbase import TextEmbeddingIndexSettings
 
 
 # TODO: RelatedTermIndexSettings belongs in relatedtermsindex.py.
